+++ ovp_projects/models/project.py
@@ -5,12 +5,10 @@ from django.template.defaultfilters import slugify
 from django.db.models import Sum
 from django.db.models.signals import post_save, post_delete
 from django.dispatch import receiver
-from ovp_core.helpers import get_settings
+from ovp_core.helpers import get_settings, get_address_model
 
-#from ovp_projects.serializers import project as serializers
-
-from ovp_projects.models.apply import Apply
 from ovp_projects import emails
+from ovp_projects.models.apply import Apply
 
 import urllib.request as request
 import urllib.parse as parse
@@ -22,7 +20,7 @@ class Project(models.Model):
   Project model
   """
   image = models.ForeignKey('ovp_uploads.UploadedImage', blank=True, null=True, verbose_name=_('image'))
-  address = models.OneToOneField('ovp_core.GoogleAddress', blank=True, null=True, verbose_name=_('address'))
+  address = models.OneToOneField(get_address_model(), blank=True, null=True, verbose_name=_('address'))
   skills = models.ManyToManyField('ovp_core.Skill', verbose_name=_('skills'))
   causes = models.ManyToManyField('ovp_core.Cause', verbose_name=_('causes'))