+++ ovp_projects/tests/runtests.py
@@ -10,6 +10,7 @@ from django.core.management import execute_from_command_line
 
 BASE_DIR = os.path.abspath(os.path.dirname(__file__))
 sys.path.insert(0, os.path.abspath(os.path.join(BASE_DIR, '../..')))
+sys.path.insert(0, os.path.abspath(os.path.join(BASE_DIR, '../../../django-ovp-core')))
 
 # Unfortunately, apps can not be installed via ``modify_settings``
 # decorator, because it would miss the database setup.
@@ -116,7 +117,7 @@ settings.configure(
     OVP_PROJECTS = {
       'CAN_CREATE_PROJECTS_WITHOUT_ORGANIZATION': True,
       'CAN_CREATE_PROJECTS_IN_ANY_ORGANIZATION': False
-    }
+    },
 )
 
 django.setup()